// packages/prisma-thinkthank/generated/zod/inputTypeSchemas/UserWhereInputSchema.ts
export const UserWhereInputSchema: z.ZodType<Prisma.UserWhereInput> = z.object({
  AND: z.union([ z.lazy(() => UserWhereInputSchema),z.lazy(() => UserWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => UserWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => UserWhereInputSchema),z.lazy(() => UserWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  name: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  email: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  emailVerified: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  image: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  isMachine: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  passwordHash: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  invalidLoginAttempts: z.union([ z.lazy(() => IntFilterSchema),z.number() ]).optional(),
  lockedAt: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  subscribed: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  source: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  defaultWorkspace: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  defaultPartnerId: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  referralLinkId: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  fullName: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  dateFormat: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  locale: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  timeFormat: z.union([ z.lazy(() => IntNullableFilterSchema),z.number() ]).optional().nullable(),
  timezone: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  weekStartsOnMonday: z.union([ z.lazy(() => BoolNullableFilterSchema),z.boolean() ]).optional().nullable(),
  accounts: z.lazy(() => AccountListRelationFilterSchema).optional(),
  sessions: z.lazy(() => SessionListRelationFilterSchema).optional(),
  projects: z.lazy(() => ProjectUsersListRelationFilterSchema).optional(),
  partners: z.lazy(() => PartnerUserListRelationFilterSchema).optional(),
  links: z.lazy(() => LinkListRelationFilterSchema).optional(),
  dashboards: z.lazy(() => DashboardListRelationFilterSchema).optional(),
  tokens: z.lazy(() => TokenListRelationFilterSchema).optional(),
  restrictedTokens: z.lazy(() => RestrictedTokenListRelationFilterSchema).optional(),
  oAuthCodes: z.lazy(() => OAuthCodeListRelationFilterSchema).optional(),
  integrations: z.lazy(() => IntegrationListRelationFilterSchema).optional(),
  installedIntegrations: z.lazy(() => InstalledIntegrationListRelationFilterSchema).optional(),
  utmTemplates: z.lazy(() => UtmTemplateListRelationFilterSchema).optional(),
  apps: z.lazy(() => AppListRelationFilterSchema).optional(),
  bankAccounts: z.lazy(() => BankAccountListRelationFilterSchema).optional(),
  Transaction: z.lazy(() => TransactionListRelationFilterSchema).optional(),
  Document: z.lazy(() => DocumentListRelationFilterSchema).optional(),
  reportAccess: z.lazy(() => ReportAccessListRelationFilterSchema).optional(),
  reports: z.lazy(() => ReportListRelationFilterSchema).optional(),
  aiAssistantSettings: z.union([ z.lazy(() => AIAssistantSettingsNullableRelationFilterSchema),z.lazy(() => AIAssistantSettingsWhereInputSchema) ]).optional().nullable(),
  financialSettings: z.union([ z.lazy(() => FinancialSettingsNullableRelationFilterSchema),z.lazy(() => FinancialSettingsWhereInputSchema) ]).optional().nullable(),
  BankConnectionSettings: z.lazy(() => BankConnectionSettingsListRelationFilterSchema).optional()
}).strict();